Created `api` folder and moved three python files
