+++ website/api/socketio_handlers.py
@@ -4,9 +4,9 @@ This code contains the main functions that communicate with the server (server s
 
 from flask import request
 from flask_login import current_user
-from .database import Player, Chat, Message
-from .utils import check_existing_chats
-from . import db
+from ..database import Player, Chat, Message
+from ..utils import check_existing_chats
+from .. import db
 from datetime import datetime
