+++ website/auth.py
@@ -16,7 +16,7 @@ from .database import Player
 from werkzeug.security import generate_password_hash, check_password_hash
 from . import db
 from flask_login import login_user, login_required, logout_user, current_user
-from website import rest_api
+from website.api import rest_api
 
 auth = Blueprint("auth", __name__)
